POC/FastAPI/requirements.txt
fastapi==0.115.5
sqlmodel==0.0.22
uvicorn==0.32.0
requests_oauthlib
python-dotenv
